Tidy up required action components
- Move RequiredAction component to activity folder
- Rename RequiredAction component to RequiredActionCard
- Rename genericComponents/RequiredActionCard to ActionCard
